doom: Hack around `org-hide-leading-stars` issue
Setting the variable to `nil` doesn't work when using the daemon,
because the value is overwritten later, after startup is completed. This
seems to be because the theme isn't loaded until the first frame is
opened, which is after daemon startup. I guess the theme is overwriting
the value in this case.

This change adds a hook that runs when `org-mode` is loaded, forcing the
value back to `nil` and toggling `font-lock-mode`, which is required to
have the new value take effect.

See issue for more info: https://github.com/hlissner/doom-emacs/issues/3076