fish: Fix undesirable `LS_COLORS` coloring
This was the reason `fzf` was showing weird colors. Turns out, `fd` was
updated to use some new defaults when `LS_COLORS` isn't defined, so it
was the source of the colors I didn't like.

This commit re-enables coloring in `fzf`, and sets `LS_COLORS` to just
use the default colors (uses the terminal's basic 16 colors).