memorize: Start Haskell version because Rust sux
Nah Rust is awesome, I just can't figure out how to cancel blocking I/O
operations.

I want to prompt the user to answer a question in a certain amount of
time. If they fail to answer in time, I want the program to continue. So
I need a way of cancelling `read_line`. This appears to be impossible in
Rust without doing unsafe things like killing threads.

Even `tokio`'s `read_line` spawns a new thread to run `std`'s
`read_line`! And if the future is cancelled, the thread gets leaked, and
the program won't terminate until it receives a newline on `stdin`.