fish: Simplify `rstoc` function using `rstoc`
The naming is confusing. Originally I had the Fish function. Inspired by
that, I made a separate project named the same. It doesn't have all the
same features yet, but it's headed in that direction, and is doing
things in a more polished and robust way. In the meantime, I'll fill in
the gaps with `fd` and `fzf` and such.